Module rental::common

Premade types for the most common use cases.

Structs

RentMut
Stores an owner and a mutable reference in the same struct.

RentMut_Borrow
Shared borrow of a rental struct.

RentMut_BorrowMut
Mutable borrow of a rental struct.

RentMutex
Stores a Mutex and a MutexGuard in the same struct.

RentMutex_Borrow
Shared borrow of a rental struct.

RentMutex_BorrowMut
Mutable borrow of a rental struct.

RentRef
Stores an owner and a shared reference in the same struct.

RentRefCell
Stores a RefCell and a Ref in the same struct.

RentRefCellMut
Stores a RefCell and a RefMut in the same struct.

RentRefCellMut_Borrow
Shared borrow of a rental struct.

RentRefCellMut_BorrowMut
Mutable borrow of a rental struct.

RentRefCell_Borrow
Shared borrow of a rental struct.

RentRefCell_BorrowMut
Mutable borrow of a rental struct.

RentRef_Borrow
Shared borrow of a rental struct.

RentRef_BorrowMut
Mutable borrow of a rental struct.

RentRwLock
Stores an RwLock and an RwLockReadGuard in the same struct.

RentRwLockMut
Stores an RwLock and an RwLockWriteGuard in the same struct.

RentRwLockMut_Borrow
Shared borrow of a rental struct.

RentRwLockMut_BorrowMut
Mutable borrow of a rental struct.

RentRwLock_Borrow
Shared borrow of a rental struct.

RentRwLock_BorrowMut
Mutable borrow of a rental struct.