Module rental::examples

Example instantiations that demonstrate the API provided by the types this crate generates.

Structs

RentMut
A simple example that can store an owner and a mutable reference in the same struct.

RentMutMap
An example that borrows mutably off of another rental struct.

RentMutMap_Borrow
Shared borrow of a rental struct.

RentMutMap_BorrowMut
Mutable borrow of a rental struct.

RentMut_Borrow
Shared borrow of a rental struct.

RentMut_BorrowMut
Mutable borrow of a rental struct.

RentRef
A simple example that can store an owner and a shared reference in the same struct.

RentRefMap
An example that borrows off of another rental struct.

RentRefMap_Borrow
Shared borrow of a rental struct.

RentRefMap_BorrowMut
Mutable borrow of a rental struct.

RentRef_Borrow
Shared borrow of a rental struct.

RentRef_BorrowMut
Mutable borrow of a rental struct.