Struct rental::examples::SimpleMut

pub struct SimpleMut { /* fields omitted */ }

The simplest mutable rental. Mutable rentals have a slightly different API; compare this struct to SimpleRef for the clearest picture of how they differ.

Implementations

impl SimpleMut

pub fn new<__Firef>(head: Box<i32>, iref: __Firef) -> Self where
__Firef: for<'head> FnOnce(&'head mut i32) -> &'head mut i32,

Create a new instance of the rental struct.

The first argument provided is the head, followed by a series of closures, one for each tail field. Each of these closures will receive, as its arguments, a borrow of the previous field, followed by borrows of the remaining prefix fields if the struct is a shared rental. If the struct is a mutable rental, only the immediately preceding field is passed.

pub fn try_new<__Firef, __E>(
head: Box<i32>,
iref: __Firef
) -> RentalResult<Self, __E, Box<i32>> where
__Firef: for<'head> FnOnce(&'head mut i32) -> Result<&'head mut i32, __E>,

Attempt to create a new instance of the rental struct.

As new, but each closure returns a Result. If one of them fails, execution is short-circuited and a tuple of the error and the original head value is returned to you.

pub fn try_new_or_drop<__Firef, __E>(
head: Box<i32>,
iref: __Firef
) -> Result<Self, __E> where
__Firef: for<'head> FnOnce(&'head mut i32) -> Result<&'head mut i32, __E>,

Attempt to create a new instance of the rental struct.

As try_new, but only the error value is returned upon failure; the head value is dropped. This method interacts more smoothly with existing error conversions.

pub unsafe fn all_erased(&self) -> Self::Borrow

Return lifetime-erased shared borrows of the fields of the struct.

This is unsafe because the erased lifetimes are fake. Use this only if absolutely necessary and be very mindful of what the true lifetimes are.

pub unsafe fn all_mut_erased(&mut self) -> Self::BorrowMut

Return a lifetime-erased mutable borrow of the suffix of the struct.

This is unsafe because the erased lifetimes are fake. Use this only if absolutely necessary and be very mindful of what the true lifetimes are.

pub fn rent<__F, __R>(&self, f: __F) -> __R where
__F: for<'iref, 'head> FnOnce(&'iref &'head mut i32) -> __R,
__R: ,

Execute a closure on the shared suffix of the struct.

The closure may return any value not bounded by one of the special rental lifetimes of the struct.

pub fn rent_mut<__F, __R>(&mut self, f: __F) -> __R where
__F: for<'iref, 'head> FnOnce(&'iref mut &'head mut i32) -> __R,
__R: ,

Execute a closure on the mutable suffix of the struct.

The closure may return any value not bounded by one of the special rental lifetimes of the struct.

pub fn ref_rent<__F, __R>(&self, f: __F) -> &__R where
__F: for<'iref, 'head> FnOnce(&'iref &'head mut i32) -> &'iref __R,
__R: ?Sized,

Return a shared reference from the shared suffix of the struct.

This is a subtle variation of rent where it is legal to return a reference bounded by a rental lifetime, because that lifetime is reborrowed away before it is returned to you.

pub fn maybe_ref_rent<__F, __R>(&self, f: __F) -> Option<&__R> where
__F: for<'iref, 'head> FnOnce(&'iref &'head mut i32) -> Option<&'iref __R>,
__R: ?Sized,

Optionally return a shared reference from the shared suffix of the struct.

This is a subtle variation of rent where it is legal to return a reference bounded by a rental lifetime, because that lifetime is reborrowed away before it is returned to you.

pub fn try_ref_rent<__F, __R, __E>(&self, f: __F) -> Result<&__R, __E> where
__F: for<'iref, 'head> FnOnce(&'iref &'head mut i32) -> Result<&'iref __R, __E>,
__R: ?Sized,

Try to return a shared reference from the shared suffix of the struct, or an error on failure.

This is a subtle variation of rent where it is legal to return a reference bounded by a rental lifetime, because that lifetime is reborrowed away before it is returned to you.

pub fn ref_rent_mut<__F, __R>(&mut self, f: __F) -> &mut __R where
__F: for<'iref, 'head> FnOnce(&'iref mut &'head mut i32) -> &'iref mut __R,
__R: ?Sized,

Return a mutable reference from the mutable suffix of the struct.

This is a subtle variation of rent_mut where it is legal to return a reference bounded by a rental lifetime, because that lifetime is reborrowed away before it is returned to you.

pub fn maybe_ref_rent_mut<__F, __R>(&mut self, f: __F) -> Option<&mut __R> where
__F: for<'iref, 'head> FnOnce(&'iref mut &'head mut i32) -> Option<&'iref mut __R>,
__R: ?Sized,

Optionally return a mutable reference from the mutable suffix of the struct.

This is a subtle variation of rent_mut where it is legal to return a reference bounded by a rental lifetime, because that lifetime is reborrowed away before it is returned to you.

pub fn try_ref_rent_mut<__F, __R, __E>(
&mut self,
f: __F
) -> Result<&mut __R, __E> where
__F: for<'iref, 'head> FnOnce(&'iref mut &'head mut i32) -> Result<&'iref mut __R, __E>,
__R: ?Sized,

Try to return a mutable reference from the mutable suffix of the struct, or an error on failure.

This is a subtle variation of rent_mut where it is legal to return a reference bounded by a rental lifetime, because that lifetime is reborrowed away before it is returned to you.

pub fn into_head(self) -> Box<i32>

Drop the rental struct and return the original head value to you.