Trait rental::IntoSuffix

pub trait IntoSuffix {
    type Suffix;
    fn into_suffix(self) -> Self::Suffix;
}

This trait converts any *_Borrow or *_BorrowMut structs generated by the rental macro into their suffix (most dependent) field.

When you own a borrow struct, such as in the body of the closure provided to the rent_all or ref_rent_all methods of a rental struct, you can call into_suffix() to discard the borrow struct and obtain the suffix field if you don't need any of the other fields.

Associated Types

type Suffix

Type of the transitive suffix of the borrow struct.

If the suffix field of the borrow struct is itself a borrow struct of a subrental, then this type is the suffix of that nested borrow struct, recursively.

Required Methods

fn into_suffix(self) -> Self::Suffix

Discard the borrow struct and return the transitive suffix field.

If the suffix field of the borrow struct is itself a borrow struct of a subrental, then this function will return the nested suffix of that borrow struct, recursively.